airflow/requirements.txt
apache-airflow==2.9.2
pandas
psycopg2-binary
pymongo
